+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0]
+
+* make LCP LQIP use new SS functionality of saving variations with different extensions. The LCPLQIP() method doesn't 
+have the format parameter anymore. Format changes can now be chained using 
+[tractorcow/silverstripe-image-formatter](https://github.com/tractorcow/silverstripe-image-formatter). 
+
 ## [2.1.5]
 
 * improve performance of LCP LQIP
@@ -22,7 +28,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.1.1]
 
-* fix LCP LQIP method, including format conversion in LQIP method because chaining is not possible for files with different extension
+* fix LCP LQIP method, including format conversion in LQIP method because chaining is not possible for files with 
+different extension
 
 ## [2.1.0]